+++ src/cc/cc_kernels.cu
@@ -287,7 +287,7 @@ __global__ void CC64_Filter_Kernel(const uint32 *partitionList,
   uint32 partition = partitionList[0];
   // uint32 touch;
   //  Start Edge
-  uint32 startEdge = d_offsets[d_partitionsOffsets[partition]];
+  uint64 startEdge = d_offsets[d_partitionsOffsets[partition]];
 
   // End Edge
   // uint32 endEdge = d_offsets[d_partitionsOffsets[partition + 1]];
@@ -308,13 +308,13 @@ __global__ void CC64_Filter_Kernel(const uint32 *partitionList,
 
     d_filterFrontier[warpIdx] = 0;
 
-    uint32 sourceValue = d_values[warpIdx];
+    uint64 sourceValue = d_values[warpIdx];
 
     const uint64 start = d_offsets[warpIdx] - startEdge;
     const uint64 end = d_offsets[warpIdx + 1] - startEdge;
 
     for (uint64 i = start + laneIdx; i < end; i += WARP_SIZE) {
-      uint32 neighborId = d_filterEdges[i];
+      uint64 neighborId = d_filterEdges[i];
 
       // If this new path has lower cost than the previous then change and add
       // the neighbor to the frontier
@@ -433,7 +433,7 @@ __global__ void CC64_NeighborFilter_Kernel(const uint32 *partitionList,
   uint32 partition = partitionList[0];
   // uint32 touch;
   //  Start Edge
-  uint32 startEdge = d_offsets[d_partitionsOffsets[partition]];
+  uint64 startEdge = d_offsets[d_partitionsOffsets[partition]];
 
   // End Edge
   // uint32 endEdge = d_offsets[d_partitionsOffsets[partition + 1]];
@@ -453,13 +453,13 @@ __global__ void CC64_NeighborFilter_Kernel(const uint32 *partitionList,
 
     d_filterFrontier[warpIdx] = 0;
 
-    uint32 sourceValue = d_values[warpIdx];
+    uint64 sourceValue = d_values[warpIdx];
 
     const uint64 start = d_offsets[warpIdx] - startEdge;
     const uint64 end = d_offsets[warpIdx + 1] - startEdge;
 
     for (uint64 i = start + laneIdx; i < end; i += WARP_SIZE) {
-      uint32 neighborId = d_filterEdges[i];
+      uint64 neighborId = d_filterEdges[i];
 
       // If this new path has lower cost than the previous then change and add
       // the neighbor to the frontier
@@ -562,13 +562,13 @@ __global__ void CC64_Static_Filter_Kernel(const uint32 *partitionList,
 
     d_filterFrontier[warpIdx] = 0;
 
-    uint32 sourceValue = d_values[warpIdx];
+    uint64 sourceValue = d_values[warpIdx];
 
     const uint64 start = d_offsets[warpIdx];
     const uint64 end = d_offsets[warpIdx + 1];
 
     for (uint64 i = start + laneIdx; i < end; i += WARP_SIZE) {
-      uint32 neighborId = d_filterEdges[i];
+      uint64 neighborId = d_filterEdges[i];
 
       // If this new path has lower cost than the previous then change and add
       // the neighbor to the frontier
