+++ src/pr/pr.cu
@@ -684,10 +684,10 @@ void PR64(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
         for (uint32 index = 0; index < partitionList.size(); index++) {
           uint32 partition = partitionList[index];
           // Partition edge start
-          uint32 start =
+          uint64 start =
               graph->h_offsets[graph->h_partitionsOffsets[partition]];
 
-          uint32 partitionSize =
+          uint64 partitionSize =
               graph->h_offsets[graph->h_partitionsOffsets[partition + 1]] -
               start;
 
@@ -720,10 +720,10 @@ void PR64(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
             partition = partitionList[index];
 
             // Partition edge start
-            uint32 neighborStart =
+            uint64 neighborStart =
                 graph->h_offsets[graph->h_partitionsOffsets[partition]];
 
-            uint32 neighborPartitionSize =
+            uint64 neighborPartitionSize =
                 graph->h_offsets[graph->h_partitionsOffsets[partition + 1]] -
                 neighborStart;
 
@@ -786,15 +786,15 @@ void PR64(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
             //       cudaDeviceSynchronize();
             uint32 processedPartition = graph->h_partitionList[tStream];
 
-            uint32 partitionStart =
+            uint64 partitionStart =
                 graph
                     ->h_offsets[graph->h_partitionsOffsets[processedPartition]];
 
-            uint32 partitionEnd =
+            uint64 partitionEnd =
                 graph->h_offsets[graph->h_partitionsOffsets[processedPartition +
                                                             1]];
 
-            uint32 processedPartitionSize = partitionEnd - partitionStart;
+            uint64 processedPartitionSize = partitionEnd - partitionStart;
 
             if (partitionEnd <= graph->numStaticEdges) {
               // cudaStreamSynchronize(staticStreams[processedPartition]);
@@ -841,15 +841,15 @@ void PR64(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
 
               uint32 processedPartition = graph->h_nPartList[gpu][nStream];
 
-              uint32 partitionStart =
+              uint64 partitionStart =
                   graph->h_offsets
                       [graph->h_partitionsOffsets[processedPartition]];
 
-              uint32 partitionEnd =
+              uint64 partitionEnd =
                   graph->h_offsets
                       [graph->h_partitionsOffsets[processedPartition + 1]];
 
-              uint32 processedPartitionSize = partitionEnd - partitionStart;
+              uint64 processedPartitionSize = partitionEnd - partitionStart;
 
               if (partitionEnd <= graph->numStaticEdges) {
                 // Aqui
@@ -910,14 +910,14 @@ void PR64(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
           //   cudaDeviceSynchronize();
           uint32 processedPartition = graph->h_partitionList[tStream];
 
-          uint32 partitionStart =
+          uint64 partitionStart =
               graph->h_offsets[graph->h_partitionsOffsets[processedPartition]];
 
-          uint32 partitionEnd =
+          uint64 partitionEnd =
               graph->h_offsets[graph->h_partitionsOffsets[processedPartition +
                                                           1]];
 
-          uint32 processedPartitionSize = partitionEnd - partitionStart;
+          uint64 processedPartitionSize = partitionEnd - partitionStart;
 
           if (partitionEnd <= graph->numStaticEdges) {
             // cudaStreamSynchronize(staticStreams[processedPartition]);
@@ -952,15 +952,15 @@ void PR64(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
 
             uint32 processedPartition = graph->h_nPartList[gpu][nStream];
 
-            uint32 partitionStart =
+            uint64 partitionStart =
                 graph
                     ->h_offsets[graph->h_partitionsOffsets[processedPartition]];
 
-            uint32 partitionEnd =
+            uint64 partitionEnd =
                 graph->h_offsets[graph->h_partitionsOffsets[processedPartition +
                                                             1]];
 
-            uint32 processedPartitionSize = partitionEnd - partitionStart;
+            uint64 processedPartitionSize = partitionEnd - partitionStart;
 
             if (partitionEnd <= graph->numStaticEdges) {
               // cudaSetDevice(gpu + 1);
@@ -1735,10 +1735,10 @@ void PR64_PUSH(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
           uint32 partition = partitionList[index];
 
           // Partition edge start
-          uint32 start =
+          uint64 start =
               graph->h_offsets[graph->h_partitionsOffsets[partition]];
 
-          uint32 partitionSize =
+          uint64 partitionSize =
               graph->h_offsets[graph->h_partitionsOffsets[partition + 1]] -
               start;
 
@@ -1770,10 +1770,10 @@ void PR64_PUSH(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
             partition = partitionList[index];
 
             // Partition edge start
-            uint32 neighborStart =
+            uint64 neighborStart =
                 graph->h_offsets[graph->h_partitionsOffsets[partition]];
 
-            uint32 neighborPartitionSize =
+            uint64 neighborPartitionSize =
                 graph->h_offsets[graph->h_partitionsOffsets[partition + 1]] -
                 neighborStart;
 
@@ -1837,15 +1837,15 @@ void PR64_PUSH(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
 
             uint32 processedPartition = graph->h_partitionList[tStream];
 
-            uint32 partitionStart =
+            uint64 partitionStart =
                 graph
                     ->h_offsets[graph->h_partitionsOffsets[processedPartition]];
 
-            uint32 partitionEnd =
+            uint64 partitionEnd =
                 graph->h_offsets[graph->h_partitionsOffsets[processedPartition +
                                                             1]];
 
-            uint32 processedPartitionSize = partitionEnd - partitionStart;
+            uint64 processedPartitionSize = partitionEnd - partitionStart;
 
             if (partitionEnd <= graph->numStaticEdges) {
               // cudaStreamSynchronize(staticStreams[processedPartition]);
@@ -1898,15 +1898,15 @@ void PR64_PUSH(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
               //     graph->d_valuesPR, graph->d_residual, graph->d_delta);
               uint32 processedPartition = graph->h_nPartList[gpu][nStream];
 
-              uint32 partitionStart =
+              uint64 partitionStart =
                   graph->h_offsets
                       [graph->h_partitionsOffsets[processedPartition]];
 
-              uint32 partitionEnd =
+              uint64 partitionEnd =
                   graph->h_offsets
                       [graph->h_partitionsOffsets[processedPartition + 1]];
 
-              uint32 processedPartitionSize = partitionEnd - partitionStart;
+              uint64 processedPartitionSize = partitionEnd - partitionStart;
 
               if (partitionEnd <= graph->numStaticEdges) {
                 // cudaSetDevice(gpu + 1);
@@ -1968,14 +1968,14 @@ void PR64_PUSH(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
 
           uint32 processedPartition = graph->h_partitionList[tStream];
 
-          uint32 partitionStart =
+          uint64 partitionStart =
               graph->h_offsets[graph->h_partitionsOffsets[processedPartition]];
 
-          uint32 partitionEnd =
+          uint64 partitionEnd =
               graph->h_offsets[graph->h_partitionsOffsets[processedPartition +
                                                           1]];
 
-          uint32 processedPartitionSize = partitionEnd - partitionStart;
+          uint64 processedPartitionSize = partitionEnd - partitionStart;
 
           if (partitionEnd <= graph->numStaticEdges) {
             // cudaStreamSynchronize(staticStreams[processedPartition]);
@@ -2013,15 +2013,15 @@ void PR64_PUSH(std::string filePath, uint32 nRuns, uint32 nNeighborGPUs,
 
             uint32 processedPartition = graph->h_nPartList[gpu][nStream];
 
-            uint32 partitionStart =
+            uint64 partitionStart =
                 graph
                     ->h_offsets[graph->h_partitionsOffsets[processedPartition]];
 
-            uint32 partitionEnd =
+            uint64 partitionEnd =
                 graph->h_offsets[graph->h_partitionsOffsets[processedPartition +
                                                             1]];
 
-            uint32 processedPartitionSize = partitionEnd - partitionStart;
+            uint64 processedPartitionSize = partitionEnd - partitionStart;
 
             if (partitionEnd <= graph->numStaticEdges) {
               cudaMemcpyAsync(&graph->d_staticEdges[partitionStart],
